pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name        = "wvh_guide"
version     = "0.1.0"
description = "Navigation guide for WVH building"
readme      = "README.md"
license     = { text = "MIT" }
authors     = [{ name="Sumer Shinde" }]
dependencies = [
  "python-dotenv",
  "numpy",
  "networkx",
  "matplotlib",
  "google-generativeai"
]

[project.scripts]
wvh-guide = "wvh_guide.__main__:main"

[tool.setuptools.packages.find]
where   = ["."]
include = ["wvh_guide*"]

[tool.setuptools.package-data]
"wvh_guide" = ["data/*.json"]
